src/requirements.txt
Flask==3.0.2
psycopg[binary]
SQLAlchemy==2.0.27
python-dotenv==1.0.1
alembic>=1.13
